+++ instrumentation/github.com/aws/aws-sdk-go-v2/otelaws/sqsattributes.go
@@ -21,7 +21,7 @@ import (
 	"github.com/aws/smithy-go/middleware"
 
 	"go.opentelemetry.io/otel/attribute"
-	semconv "go.opentelemetry.io/otel/semconv/v1.17.0"
+	semconv "go.opentelemetry.io/otel/semconv/v1.21.0"
 )
 
 // SQSAttributeSetter sets SQS specific attributes depending on the SQS operation being performed.